Pull registration functionality out of BlockTypeIds
