+++ src/block/BlockIdentifier.php
@@ -36,7 +36,7 @@ class BlockIdentifier{
 		private string $blockTypeId,
 		private ?string $tileClass = null
 	){
-		$this->typeNumber = BlockTypeIds::lookupTypeNumberFromTypeId($this->blockTypeId);
+		$this->typeNumber = self::lookupTypeNumberFromTypeId($this->blockTypeId);
 		if($tileClass !== null){
 			Utils::testValidInstance($tileClass, Tile::class);
 		}
@@ -54,5 +54,53 @@ class BlockIdentifier{
 	 */
 	public function getTileClass() : ?string{
 		return $this->tileClass;
+	}
+
+	public const AIR_TYPE_NUMBER = 10000;
+
+	private static int $nextTypeNumber = self::AIR_TYPE_NUMBER + 1; //fixed ID reserved for air, for Block::EMPTY_STATE_ID
+
+	/**
+	 * @var int[]
+	 * @phpstan-var array<string, int>
+	 */
+	private static $typeIdToTypeNumber = [];
+	/**
+	 * @var string[]
+	 * @phpstan-var array<int, string>
+	 */
+	private static $typeNumberToTypeId = [];
+
+	/**
+	 * @var int[]
+	 * @phpstan-var array<int, int>
+	 */
+	private static $typeIdXorMasks = [];
+
+	public static function firstUnusedTypeNumber() : int{
+		return self::$nextTypeNumber;
+	}
+
+	private static function claimTypeId(string $typeId) : int{
+		if(isset(self::$typeIdToTypeNumber[$typeId])){
+			throw new \InvalidArgumentException("Type ID \"$typeId\" has already been claimed");
+		}
+		$typeNumber = $typeId === BlockTypeIds::AIR ? self::AIR_TYPE_NUMBER : self::$nextTypeNumber++;
+		self::$typeIdToTypeNumber[$typeId] = $typeNumber;
+		self::$typeNumberToTypeId[$typeNumber] = $typeId;
+		self::$typeIdXorMasks[$typeNumber] = Block::computeStateIdXorMask($typeNumber);
+		return $typeNumber;
+	}
+
+	public static function lookupTypeNumberFromTypeId(string $typeId) : int{
+		return self::$typeIdToTypeNumber[$typeId] ??= self::claimTypeId($typeId);
+	}
+
+	public static function stateIdXorMask(int $typeId) : int{
+		return self::$typeIdXorMasks[$typeId];
+	}
+
+	public static function lookupTypeIdFromTypeNumber(int $typeNumber) : string{
+		return self::$typeNumberToTypeId[$typeNumber] ?? throw new \InvalidArgumentException("Unknown type number $typeNumber (probably not registered on this thread?)");
 	}
 }
